Code Commit @Surendra
Auto Increment Common Id Genrator
