+++ src/main/java/com/hrms/repository/DepartmentDao.java
@@ -9,6 +9,6 @@ import com.hrms.model.Department;
 
 public interface DepartmentDao extends GenericDao<Department>
 {
-	String getMAX_DEPARTMENT_CODE();
+
 	
 }
