+++ src/main/java/com/hrms/repository/GradeMaterDao.java
@@ -5,6 +5,6 @@ import com.hrms.model.Grade;
 
 public interface GradeMaterDao extends GenericDao<Grade> {
 
-	int getMAX_GRADE_CODE();
+
 
 }
